ai_gemini_microservice/requirements.txt
flask>=2.2
requests>=2.28
# Optional: to enable real Gemini calls, the official Google Generative AI client
# Install in the container by including the package below. Note: pass a valid
# GOOGLE_API_KEY when you run the container.
google-generativeai==0.8.3

